feat: finalisasi proyek

+++ database/migrations/2025_07_31_083339_add_user_id_to_banners_table_final.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('banners', function (Blueprint $table) {
-            $table->foreignId('user_id')->nullable()->constrained()->onDelete('cascade');
+            if (!Schema::hasColumn('banners', 'user_id')) {
+                $table->foreignId('user_id')->nullable()->constrained()->onDelete('cascade');
+            }
         });
     }
